+++ bench_method.go
@@ -27,6 +27,8 @@ import (
 
 	"github.com/yarpc/yab/encoding"
 	"github.com/yarpc/yab/transport"
+
+	"github.com/opentracing/opentracing-go"
 )
 
 type benchmarkMethod struct {
@@ -37,8 +39,7 @@ type benchmarkMethod struct {
 // WarmTransport warms up a transport and returns it. The transport is warmed
 // up by making some number of requests through it.
 func (m benchmarkMethod) WarmTransport(opts TransportOptions, warmupRequests int) (transport.Transport, error) {
-	opts.benchmarking = true
-	transport, err := getTransport(opts, m.serializer.Encoding())
+	transport, err := getTransport(opts, m.serializer.Encoding(), opentracing.NoopTracer{})
 	if err != nil {
 		return nil, err
 	}