+++ transport/tchannel.go
@@ -28,6 +28,7 @@ import (
 	"io/ioutil"
 	"os"
 
+	"github.com/opentracing/opentracing-go"
 	"github.com/uber/tchannel-go"
 	"github.com/uber/tchannel-go/thrift"
 	"golang.org/x/net/context"
@@ -40,6 +41,7 @@ const rawHeadersKey = "_raw_"
 type tchan struct {
 	sc          *tchannel.SubChannel
 	callOptions *tchannel.CallOptions
+	tracer      opentracing.Tracer
 }
 
 // TChannelOptions are used to create a TChannel transport.
@@ -49,6 +51,19 @@ type TChannelOptions struct {
 
 	// TargetService is the service name being targeted.
 	TargetService string
+
+	// RoutingDelegate is a traffic group that overrides the routing key,
+	// to redirect to an application layer traffic proxy.
+	RoutingDelegate string
+
+	// RoutingKey is a traffic group that overrides the service name, for a
+	// proxy to redirect to a more specific traffic group than the service
+	// proper.
+	RoutingKey string
+
+	// ShardKey is an opaque blob that clues where to direct a request to an
+	// instance within a traffic group.
+	ShardKey string
 
 	// LogLevel overrides the default LogLevel (Warn).
 	LogLevel *tchannel.LogLevel
@@ -65,6 +80,10 @@ type TChannelOptions struct {
 	// TransportOpts are a list of options, mostly used to add or override
 	// TChannel's transport headers.
 	TransportOpts map[string]string
+
+	// Tracer is an instance of an opentracing tracer for baggage propagation
+	// and/or span submission.
+	Tracer opentracing.Tracer
 }
 
 // NewTChannel returns a Transport that calls a TChannel service.
@@ -85,11 +104,10 @@ func NewTChannel(opts TChannelOptions) (Transport, error) {
 	}
 	processName := fmt.Sprintf("%v@%v:%v[%v]", os.Getenv("USER"), hostname, os.Args[0], os.Getpid())
 
-	// TODO: set trace sample rate to 1 for the initial request.
 	ch, err := tchannel.NewChannel(callerName, &tchannel.ChannelOptions{
-		Logger:          tchannel.NewLevelLogger(tchannel.SimpleLogger, level),
-		ProcessName:     processName,
-		TraceSampleRate: &opts.TraceSampleRate,
+		Logger:      tchannel.NewLevelLogger(tchannel.SimpleLogger, level),
+		ProcessName: processName,
+		Tracer:      opts.Tracer,
 	})
 	if err != nil {
 		return nil, fmt.Errorf("failed to create TChannel: %v", err)
@@ -102,12 +120,18 @@ func NewTChannel(opts TChannelOptions) (Transport, error) {
 	callOpts := &tchannel.CallOptions{
 		Format: tchannel.Format(opts.Encoding),
 	}
+	applyRPCOptions(callOpts, opts)
 	applyTChanOptions(callOpts, opts.TransportOpts)
 
 	return &tchan{
 		sc:          ch.GetSubChannel(opts.TargetService),
 		callOptions: callOpts,
+		tracer:      opts.Tracer,
 	}, nil
+}
+
+func (t *tchan) Tracer() opentracing.Tracer {
+	return t.tracer
 }
 
 func (t *tchan) Protocol() Protocol {
@@ -115,12 +139,23 @@ func (t *tchan) Protocol() Protocol {
 }
 
 func (t *tchan) Call(ctx context.Context, r *Request) (*Response, error) {
-	call, err := t.sc.BeginCall(ctx, r.Method, t.callOptions)
+	// We must create a shallow copy of the request headers because, at time of
+	// writing, we cannot prepare the trace headers before obtaining a TChannel
+	// call object. Consequently, we have to inject the headers and alter the
+	// request object for every call in a benchmark. Creating a shallow copy of
+	// the request object allows us to overwrite the headers reference without
+	// introducing a data race.
+	req := *r
+
+	call, err := t.sc.BeginCall(ctx, req.Method, t.callOptions)
+
 	if err != nil {
 		return nil, fmt.Errorf("begin call failed: %v", err)
 	}
 
-	if err := t.writeArgs(call, r); err != nil {
+	req.Headers = tchannel.InjectOutboundSpan(call.Response(), req.Headers)
+
+	if err := t.writeArgs(call, &req); err != nil {
 		return nil, err
 	}
 
@@ -129,8 +164,8 @@ func (t *tchan) Call(ctx context.Context, r *Request) (*Response, error) {
 		return nil, err
 	}
 
-	span := tchannel.CurrentSpan(ctx)
-	res.TransportFields["trace"] = fmt.Sprintf("%x", span.TraceID())
+	tchSpan := tchannel.CurrentSpan(ctx)
+	res.TransportFields["trace"] = fmt.Sprintf("%x", tchSpan.TraceID())
 	return res, nil
 }
 
@@ -216,12 +251,21 @@ func (t *tchan) writeArgs(call *tchannel.OutboundCall, r *Request) error {
 	return nil
 }
 
+func applyRPCOptions(callOpts *tchannel.CallOptions, opts TChannelOptions) {
+	callOpts.RoutingDelegate = opts.RoutingDelegate
+	callOpts.RoutingKey = opts.RoutingKey
+	callOpts.ShardKey = opts.ShardKey
+}
+
 func applyTChanOptions(callOpts *tchannel.CallOptions, opts map[string]string) {
 	if format, ok := opts["as"]; ok {
 		callOpts.Format = tchannel.Format(format)
 	}
 	if rd, ok := opts["rd"]; ok {
 		callOpts.RoutingDelegate = rd
+	}
+	if rk, ok := opts["rk"]; ok {
+		callOpts.RoutingKey = rk
 	}
 	if sk, ok := opts["sk"]; ok {
 		callOpts.ShardKey = sk