+++ utils_for_test.go
@@ -23,11 +23,15 @@ package main
 import (
 	"bytes"
 	"fmt"
+	"io"
 	"io/ioutil"
 	"runtime"
 	"testing"
 
+	"github.com/opentracing/opentracing-go"
 	"github.com/stretchr/testify/require"
+	"github.com/uber/jaeger-client-go"
+	jaeger_config "github.com/uber/jaeger-client-go/config"
 )
 
 // Constants useful for tests
@@ -66,4 +70,9 @@ func writeFile(t *testing.T, prefix, contents string) string {
 	require.NoError(t, err, "Write to temp file failed")
 	require.NoError(t, f.Close(), "Close temp file failed")
 	return f.Name()
+}
+
+func getTestTracer(serviceName string) (opentracing.Tracer, io.Closer, error) {
+	var jaegerConfig jaeger_config.Configuration
+	return jaegerConfig.New(serviceName, jaeger.NullStatsReporter)
 }